% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNetIterations.R
\name{NNetIterations}
\alias{NNetIterations}
\title{Title}
\usage{
NNetIterations(X.mat, y.vec, max.iterations, step.size, n.hidden.units,
  is.train)
}
\arguments{
\item{X.mat}{}

\item{y.vec}{}

\item{max.iterations}{}

\item{step.size}{}

\item{n.hidden.units}{}

\item{is.train}{}
}
\description{
Title
}
\examples{

data(ozone, package="ElemStatLearn")
X.mat <- as.matrix(ozone[,-1])
y.vec <- as.vector(ozone[,1])
n.hidden.units <- 5
max.iterations <- 50
is.train <- TRUE
step.size <- 0.01
res <- NNetIterations( X.mat, y.vec, max.iterations, step.size, n.hidden.units, is.train)

}
